UI <- dev

+++ Assets/_Project/01_Scripts/2.Controllers/BaseMonsterDate.cs
@@ -12,10 +12,10 @@ public class BaseMonsterData : ScriptableObject
     public float monsterScale = 10f;
 
     [Header("공통 전투 데이터")]
-    public float maxHp; //최대체력
+    public double maxHp; //최대체력
     public float attackRange; //몬스터 거리
 
     [Header("공통 보상")]
-    public int dropGold = 100; //보상골드
+    public double dropGold = 100; //보상골드
     public GameObject goldPrefab;
 }
